feat(plan): /autobot:plan + Phase 2.5 미리보기 신설 — 기획·디자인 검토 게이트
mvp 의 자율 흐름은 architect/ux-designer 의 첫 패스 결과를 사람 검토 없이
Phase 3–5 의 코드로 변환한다. 잘못된 기획·디자인 위에 5분간 코드를 만들고
나서야 발견하면 모두 폐기. /autobot:plan 은 그 비용을 0 으로 떨어뜨린다 —
Phase 0–2 까지 빌드하고 Phase 2.5 에서 정지, designs/preview/index.html
(self-contained, 모바일 갤러리 + 기획 요약 + nav flow + token swatch +
critique) 을 브라우저로 자동 표면화. OK 면 /autobot:resume 으로 Phase 3.

신규
- commands/plan.md — 진입점. dispatcher 의 manual-skip override 가 prose
  reconciliation 이 아닌 7-step 명시 enumeration (start-phase 2.5 → 스킬
  → advance → build-lock release → STOP)
- skills/autobot-plan-preview/ — Phase 2.5 self 단계. 멀티모달 critique
  작성 (축 1 기획 정합성, 축 2 디자인/HIG; safe area 침범이 1순위)
- scripts/build_preview.py + build-preview.sh — self-contained HTML 빌더
  (외부 CDN 0, base64 PNG, iPhone aspect 393/852 깨끗한 베젤 frame)

spec
- phases."2.5" + gates."2.5->3" 추가. manual:true 로 mvp 자율 흐름은
  자동 skip — 전용 명령 /autobot:plan 이 명시 트리거
- Phase 3.dependencies 는 ["2"] 유지 — 2.5 는 옵셔널 사이드 노드

호환성 (advisor 가 BLOCKING 으로 잡은 phase id 정렬 안전성)
- scripts/event_log.py — int(phase) 가 "2.5" 에서 ValueError. 정수 케이스는
  back-compat 으로 int 저장, 비정수는 str 로 저장하도록 lenient 처리
- scripts/render_pipeline_docs.py — int(phase_id) 정렬을 float() 로 일반화

문서
- skills/autobot-orchestrator/SKILL.md — dispatcher 결정 로직에 manual
  phase 자동 skip + 전용 명령 트리거 명시 (auto-rendered phase summary
  도 2.5 행 갱신)
- skills/autobot-orchestrator/references/phase-gates.md — Gate 2.5→3 추가
  (file_exists 만 검사, critique 품질 강제는 스킬 contract)
- commands/resume.md — phase 2.5 행 + Phase 3 의 "2.5 pending 무방" 명시
- README.md — auto-rendered

Stitch prompt 강화 (root cause fix, 다음 빌드부터 효과)
- skills/autobot-ux-design/SKILL.md — Layout Requirements 를 픽셀 단위
  3-zone 표 (status 0–47pt / content 47–818pt / home indicator 818–852pt)
  + 시각 chrome 강제 (9:41 시계, 신호/wifi/배터리, home indicator pill)
  로 교체. Stitch image generator 가 chrome 자체를 그려야 safe area 가
  자연스럽게 보호된다. 침범 시 Phase 2.5 critique HIGH severity →
  /autobot:resume 2 --force 재생성 안내

dry-run: SnapCart (LabelCatch Stitch project, 6 화면) 에서 build-preview.sh
단독 OK — self-contained HTML, 6 모바일 카드 + 5 swatch + 6 typography,
critique placeholder 정상.

+++ scripts/render_pipeline_docs.py
@@ -19,7 +19,8 @@ def load_spec() -> dict:
 
 
 def sorted_phases(spec: dict) -> list[tuple[str, dict]]:
-    return sorted(spec.get("phases", {}).items(), key=lambda item: int(item[0]))
+    # float() — supports fractional phase ids ("2.5") in addition to integer ones.
+    return sorted(spec.get("phases", {}).items(), key=lambda item: float(item[0]))
 
 
 def gate_label(gate_id: str | None) -> str:
@@ -45,7 +46,7 @@ def render_readme_phase_table(spec: dict) -> str:
 
 def render_readme_gate_summary(spec: dict) -> str:
     lines = []
-    for gate_id, gate in sorted(spec.get("gates", {}).items(), key=lambda item: int(item[1]["fromPhase"])):
+    for gate_id, gate in sorted(spec.get("gates", {}).items(), key=lambda item: float(item[1]["fromPhase"])):
         label = gate_label(gate_id)
         summary = gate.get("docsSummary", gate.get("summary", "")).strip()
         suffix = " (soft gate)" if gate.get("soft") else ""
